requirements.txt
google-generativeai
python-dotenv
langchain-core
langchain-google-genai
duckduckgo-search
pypdf
fastapi
uvicorn
python-multipart
pandas
scikit-learn
lightgbm
joblib
anyio
pip-system-certs
groq
